modules: app: Move shadow_get() to transport module
The shadow_get() function is moved to the transport module
to simplify the app module.
